Move ConversionReportingNotificationCTAWidget to GA4.
