+++ assets/js/modules/analytics-4/components/widgets/ConversionReportingNotificationCTAWidget.test.js
@@ -37,16 +37,16 @@ import {
 	KM_ANALYTICS_TOP_TRAFFIC_SOURCE_DRIVING_LEADS,
 	KM_ANALYTICS_TOP_TRAFFIC_SOURCE_DRIVING_PURCHASES,
 	KM_SEARCH_CONSOLE_POPULAR_KEYWORDS,
-} from '../../googlesitekit/datastore/user/constants';
+} from '../../../../googlesitekit/datastore/user/constants';
 import {
 	MODULES_ANALYTICS_4,
 	ENUM_CONVERSION_EVENTS,
-} from '../../modules/analytics-4/datastore/constants';
-import { KEY_METRICS_SELECTION_PANEL_OPENED_KEY } from './constants';
-import { CORE_UI } from '../../googlesitekit/datastore/ui/constants';
-import { CORE_SITE } from '../../googlesitekit/datastore/site/constants';
-import * as tracking from '../../util/tracking';
-import { getWidgetComponentProps } from '../../googlesitekit/widgets/util';
+} from '../../datastore/constants';
+import { KEY_METRICS_SELECTION_PANEL_OPENED_KEY } from '../../../../components/KeyMetrics/constants';
+import { CORE_UI } from '../../../../googlesitekit/datastore/ui/constants';
+import { CORE_SITE } from '../../../../googlesitekit/datastore/site/constants';
+import * as tracking from '../../../../util/tracking';
+import { getWidgetComponentProps } from '../../../../googlesitekit/widgets/util';
 import {
 	render,
 	createTestRegistry,
@@ -57,10 +57,10 @@ import {
 	act,
 	fireEvent,
 	provideKeyMetrics,
-} from '../../../../tests/js/test-utils';
+} from '../../../../../../tests/js/test-utils';
 import ConversionReportingNotificationCTAWidget from './ConversionReportingNotificationCTAWidget';
-import { enabledFeatures } from '../../features';
-import { VIEW_CONTEXT_MAIN_DASHBOARD } from '../../googlesitekit/constants';
+import { enabledFeatures } from '../../../../features';
+import { VIEW_CONTEXT_MAIN_DASHBOARD } from '../../../../googlesitekit/constants';
 
 const mockTrackEvent = jest.spyOn( tracking, 'trackEvent' );
 mockTrackEvent.mockImplementation( () => Promise.resolve() );
