+++ assets/js/modules/analytics-4/index.js
@@ -125,7 +125,7 @@ import GoogleTagIDMismatchNotification from './components/notifications/GoogleTa
 import { isValidPropertyID, isValidWebDataStreamID } from './utils/validation';
 import { LEGACY_ENHANCED_MEASUREMENT_ACTIVATION_BANNER_DISMISSED_ITEM_KEY } from './constants';
 import { PRIORITY } from '../../googlesitekit/notifications/constants';
-import ConversionReportingNotificationCTAWidget from '../../components/KeyMetrics/ConversionReportingNotificationCTAWidget';
+import ConversionReportingNotificationCTAWidget from './components/widgets/ConversionReportingNotificationCTAWidget';
 import { isFeatureEnabled } from '../../features';
 
 export { registerStore } from './datastore';
